message-model & get profile & get message TTL index

+++ src/utils/error/index.js
@@ -1,6 +1,6 @@
 
 import { Token } from "../../db/model/token.model.js";
-import { verifyToken } from "../token/index.js";
+import { generateToken, verifyToken } from "../token/index.js";
 
 export default function asyncHadler(fn) {
     return (req, res, next) => {
@@ -11,8 +11,13 @@ export default function asyncHadler(fn) {
 }
 
 export const globalError = async (err, req, res, next) => {
+try {
+    
     if (err.message == "jwt expired") {
-        const refreshToken = req.headers.refreshToken;
+        const refreshToken = req.headers.refreshtoken;
+        if (!refreshToken) {
+            throw new Error("Refresh token is required", { cause: 401 })
+        }
         const payload = verifyToken(refreshToken)
 
         const tokenExist = await Token.findOneAndDelete({
@@ -23,33 +28,38 @@ export const globalError = async (err, req, res, next) => {
         if (!tokenExist) {
             throw new Error("Unauthorized", { cause: 401 })
         }
-       
-            const accessToken = generateToken({ 
-                payload: { id: payload.id },
-                options:{expiresIn:"15m"}
-                
-            })
 
-            const newRefreshToken = generateToken({ 
-                payload: { id: payload.id },
-                options:{expiresIn:"7d"}
-                
-            })
+        const accessToken = generateToken({
+            payload: { id: payload.id },
+            options: { expiresIn: "15m" }
 
-            await Token.create({
-                token:newRefreshToken,
-                user:payload.id,
-                type:"refresh"
-            })
+        })
 
+        const newRefreshToken = generateToken({
+            payload: { id: payload.id },
+            options: { expiresIn: "7d" }
 
-            return res.status(200).json({ message: "User logged in successfully", success: true, accessToken,newRefreshToken })
-        
+        })
+
+        await Token.create({
+            token: newRefreshToken,
+            user: payload.id,
+            type: "refresh",
+            expiresAt: new Date(Date.now() + 7 * 24 * 60 * 60 * 1000)
+        })
+
+        return res.status(200).json({ message: "Tokens refreshed successfully", success: true, accessToken, newRefreshToken })
+
     }
     if (req.file) {
         fs.unlinkSync(req.file.path)
     }
+} catch (error) {
+    
     return res.status(err.cause || 500).json({
         message: err.message, success: false, stack: err.stack
     })
 }
+
+} 
+
